Update DataReader.java
This data reader should work properly and read in all the data as formatted in the file. The problem with previous iterations was the boundary condition when the last few entries were empty strings "", which would appear in the file as ,,,,,,,, so the scanner would ignore some of the strings and leave some of the last entries of the array null.